Revert "Revert "Several changes""
This reverts commit a8e4b11051a64230e2cd2ebb91b4654173004b69.
